Added highscore
This system gets the top 10 highscores from a file.
If no file exists, it makes a new one.
At a game over or game win, the top highscore of the player(s) gets added to the file

+++ DigDug/Wall.cpp
@@ -104,7 +104,7 @@ void game::WallComponent::RemoveSide(Directions direction)
 
 void game::WallComponent::DigWalls(const dae::ColliderComponent* playerCollider)
 {
-	for (int i{}; i < m_Colliders.size(); ++i)
+	for (size_t i{}; i < m_Colliders.size(); ++i)
 	{
 		if (!m_Colliders[i]->IsColliding(playerCollider))
 		{
